requirements.txt
bson==0.5.10
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.17.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
